requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
Werkzeug==2.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.7 